+++ src/Quartz.Serialization.SystemTextJson/Converters/CalendarConverter.cs
@@ -105,7 +105,7 @@ internal sealed class CalendarConverter : JsonConverter<ICalendar>
     {
         if (string.IsNullOrWhiteSpace(typeName) || !converters.TryGetValue(typeName!, out var converter))
         {
-            throw new ArgumentException("Don't know how to handle " + typeName);
+            throw new ArgumentException("Don't know how to handle " + typeName, nameof(typeName));
         }
 
         return converter;